cache_me_if_you_can/session_demo/requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
redis==5.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
